dois modulos

+++ lib/helpers/products_class.dart
@@ -1,10 +1,8 @@
-import 'dart:ffi';
-
 class Product {
   int id;
   String title;
   String description;
-  Float price;
+  int price;
   String imagemUrl;
 
   Product({this.id, this.title, this.description, this.price, this.imagemUrl});
